Moving mdx-components to src/components
