--- src/components/dev-dot-content/mdx-components/index.tsx
@@ -1,9 +1,0 @@
-export * from './mdx-blockquote'
-export * from './mdx-headings'
-export * from './mdx-lists'
-export * from './mdx-tabs'
-export * from './mdx-table'
-export * from './mdx-a'
-export * from './mdx-p'
-export * from './mdx-inline-code'
-export * from './mdx-hcp-callout'
