--- src/components/dev-dot-content/mdx-components/mdx-tabs/mdx-tabs.module.css
@@ -1,3 +1,0 @@
-.tabsWrapper {
-  margin: 24px 0;
-}
