+++ src/components/mdx-components/index.tsx
@@ -1,0 +1,9 @@
+export * from './mdx-a'
+export * from './mdx-blockquote'
+export * from './mdx-hcp-callout'
+export * from './mdx-headings'
+export * from './mdx-inline-code'
+export * from './mdx-lists'
+export * from './mdx-p'
+export * from './mdx-table'
+export * from './mdx-tabs'
