+++ src/components/mdx-components/mdx-tabs/mdx-tabs.module.css
@@ -1,0 +1,3 @@
+.tabsWrapper {
+	margin: 24px 0;
+}
