+++ src/views/product-root-docs-path-landing/components/paragraph-block/index.tsx
@@ -1,5 +1,5 @@
 import { MDXRemote } from 'next-mdx-remote'
-import { MdxA, MdxP } from 'components/dev-dot-content/mdx-components'
+import { MdxA, MdxP } from 'components/mdx-components'
 
 export function ParagraphBlock(props) {
 	return <MDXRemote {...props.mdxSource} components={{ a: MdxA, p: MdxP }} />
